refactor: decompose index file
